Modularize code by separating concerns
