+++ src/modules/add_tasks.js
@@ -1,21 +1,19 @@
 import tasksData from './local_data.js';
 import { listContainer, loadTasks } from './load.js';
 
-const taskToAdd = document.querySelector('#add');
-
-const addTask = () => {
+const addTask = (taskInput) => {
   const tasks = tasksData.fetchData();
-  if (taskToAdd.value !== '') {
-    const todo = { description: taskToAdd.value, completed: false, id: tasks.length };
+  if (taskInput.value !== '') {
+    const todo = { description: taskInput.value, completed: false, id: tasks.length };
     listContainer.innerHTML += `
     <li>
       <input type="checkbox" name="task" value="task1" class="to-do">
-      <span class="task-description">${taskToAdd.value}<span>
+      <span class="task-description">${taskInput.value}<span>
       <i class="fas fa-ellipsis-v option"></i>
       <i class="fas fa-trash-alt delete" id="${tasks.id}></i>
    </li>
     `;
-    taskToAdd.value = '';
+    taskInput.value = '';
     tasks.push(todo);
     tasksData.setData(tasks);
     listContainer.innerHTML = '';
