+++ src/modules/checkbox.js
