+++ src/modules/update_indexes.js
